Update 2016-02-22-private-classes-in-ruby.md
I was curious to know whether you still think that private classes are a good idea. One of my colleagues came across this blog post and I was honestly pretty surprised to see it because I've seen other very sensible things on the Arkency blog. I cannot think of any sensible reason for doing what is described here - either a) add a comment stating that this class is just an implementation detail, or b) don't document this class in your application's documentation.

I would always recommend putting a class in a separate file so that it's easy to find, can grow as big as it needs to be, and to keep your codebase nice and maintainable. It also saves you time later if you decide that you do, after all, want to use this class elsewhere. I can see almost no advantage to doing what you suggest here but plenty of disadvantages.

I hope you don't take my suggested changes to heart. :-)

Cheers